fix(reborn): address pending gate review feedback (#3867)
